+++ ql/termstructures/volatility/equityfx/Makefile.am
@@ -8,14 +8,10 @@ this_include_HEADERS = \
     andreasenhugevolatilityinterpl.hpp \
     andreasenhugevolatilityadapter.hpp \
     blackconstantvol.hpp \
-    blackinvertedvoltermstructure.hpp \
     blackvariancecurve.hpp \
     blackvariancesurface.hpp \
     blackvoltermstructure.hpp \
     fixedlocalvolsurface.hpp \
-    fxblackvolsurface.hpp \
-    fxsmilesection.hpp \
-    fxvannavolgasmilesection.hpp \
     gridmodellocalvolsurface.hpp \
     hestonblackvolsurface.hpp \
     impliedvoltermstructure.hpp \
@@ -33,8 +29,6 @@ cpp_files = \
     blackvariancesurface.cpp \
     blackvoltermstructure.cpp \
     fixedlocalvolsurface.cpp \
-    fxblackvolsurface.cpp \
-    fxvannavolgasmilesection.cpp \
     gridmodellocalvolsurface.cpp \
     hestonblackvolsurface.cpp \
     localvolsurface.cpp \
